School/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.10
redis==5.0.1
celery==5.3.4
django-redis==5.4.0
Pillow==10.2.0
django-filter==23.5
boto3==1.34.0
botocore==1.34.0
djangorestframework-camel-case==1.4.2
drf-yasg==1.21.7
python-decouple==3.8
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
channels==4.0.0
channels-redis==4.2.0
django-storages==1.14.2
openai==1.12.0
pandas==2.1.4
scikit-learn==1.3.2
uvicorn==0.27.0
websockets==12.0
django-health-check==3.18.1
sentry-sdk==1.40.0
django-prometheus==2.3.1
dj-database-url==3.0.1

# Barcode generation
python-barcode==0.15.1
django-otp==1.5.0
qrcode==7.4.2
django-two-factor-auth==1.16.0

# HTTP client used by ai_service_client
requests==2.32.3
